refactor: 인수테스트 영역으로 추가 및 파라미터는 map으로 넣기
